# front-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OOUN Restaurant API Documentation",
        "description": "University Restaurant Management System API - OOUN Tunisia",
        "contact": {
            "name": "OOUN Restaurant System",
            "email": "admin@ooun.tn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/admin/feedback": {
            "get": {
                "tags": [
                    "Admin - Feedback Management"
                ],
                "summary": "Get all feedback",
                "description": "Retrieve paginated list of feedback with filtering options",
                "operationId": "2e63332af1e7c73035ebe06c02963a1a",
                "parameters": [
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter by rating (1-5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 5,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "taste",
                                "quality",
                                "service",
                                "cleanliness"
                            ]
                        }
                    },
                    {
                        "name": "sentiment",
                        "in": "query",
                        "description": "Filter by sentiment",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "positive",
                                "neutral",
                                "negative"
                            ]
                        }
                    },
                    {
                        "name": "menu_id",
                        "in": "query",
                        "description": "Filter by menu ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "Filter by student ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/feedback/{id}": {
            "get": {
                "tags": [
                    "Admin - Feedback Management"
                ],
                "summary": "Get specific feedback",
                "description": "Retrieve detailed information about a feedback entry",
                "operationId": "bfb7512dfce5da9dfaff34544e732b81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Feedback not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Feedback Management"
                ],
                "summary": "Delete feedback",
                "description": "Remove a feedback entry from the system",
                "operationId": "dabc993d7331567f62d4b87b19b8fcb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Feedback not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/feedback/statistics": {
            "get": {
                "tags": [
                    "Admin - Statistics"
                ],
                "summary": "Get feedback statistics",
                "description": "Retrieve comprehensive feedback statistics including averages and distribution",
                "operationId": "17d9181c45aaa5cfb08ef31741024e5c",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for statistics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for statistics (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_feedback": {
                                                    "type": "integer"
                                                },
                                                "average_rating": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "by_rating": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "by_category": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "by_sentiment": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}